added link to nameartists and started styling hover effect on album
