--- src/ArtistName.jsx
@@ -1,9 +1,0 @@
-import "./ArtistName.css";
-
-export const ArtistName = ({ artists }) => {
-  return (
-    <span className="artistName">
-      {artists.map((artist) => artist.name).join(", ")}
-    </span>
-  );
-};
