--- src/CoverImage.css
@@ -1,24 +1,0 @@
-.card {
-  display: flex;
-}
-
-.cover-image {
-  width: 100%;
-  height: auto;
-}
-
-.icons {
-  display: flex;
-  justify-content: center;
-  width: 100%;
-  height: 50px;
-}
-
-#icon-small {
-  width: 35px;
-  padding: 5px;
-}
-
-#icon-big {
-  width: 60px;
-}
