+++ src/assets/Components/CoverImage/CoverImage.css
@@ -1,0 +1,64 @@
+.container {
+  position: relative;
+  width: 100%;
+}
+.cover-image {
+  width: 100%;
+  height: auto;
+}
+
+/* .icons {
+  position: absolute;
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  height: 50px;
+} */
+.icons {
+  display: flex;
+  justify-content: center;
+  width: 100%;
+  height: 50px;
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  height: 100%;
+  width: 100%;
+  opacity: 0;
+  transition: 0.3s ease;
+  background-color: rgba(245, 245, 245, 0.274);
+}
+
+#icon-small {
+  width: 35px;
+  padding: 5px;
+}
+
+#icon-big {
+  width: 60px;
+}
+
+.container:hover .icons {
+  opacity: 1;
+}
+
+#icon-small {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+}
+#icon-big {
+  color: white;
+  position: absolute;
+  top: 50%;
+  left: 50%;
+}
+/* to change color because are black , color property doesn't work */
+#icon-big:hover {
+  filter: invert(100%);
+}
+#icon-small:hover {
+  filter: invert(100%);
+}
